python/standard/vendoring/requirements.txt
google-cloud==0.20.0